Renamed folder and added to RibbonTab
Renamed folder from "IDCRevit" to "BackgroundTasks"
Added a new event on synchronizing with central to the RibbonTab project that runs the method to create a global parameter called "Add ALL Sheets to HOK Print Set" and add all sheets to a print set called "_HOK Automated Print Set"
